refactor: migrate stocks and options decimals

+++ src/options/model.rs
@@ -1,25 +1,25 @@
-use crate::common::time::Timestamp;
+use crate::common::{decimal::Decimal, time::Timestamp};
 
 #[derive(Clone, Debug, Default, PartialEq, serde::Deserialize)]
 pub struct Bar {
     pub t: Option<Timestamp>,
-    pub o: Option<f64>,
-    pub h: Option<f64>,
-    pub l: Option<f64>,
-    pub c: Option<f64>,
+    pub o: Option<Decimal>,
+    pub h: Option<Decimal>,
+    pub l: Option<Decimal>,
+    pub c: Option<Decimal>,
     pub v: Option<u64>,
     pub n: Option<u64>,
-    pub vw: Option<f64>,
+    pub vw: Option<Decimal>,
 }
 
 #[derive(Clone, Debug, Default, PartialEq, serde::Deserialize)]
 pub struct Quote {
     pub t: Option<Timestamp>,
     pub bx: Option<String>,
-    pub bp: Option<f64>,
+    pub bp: Option<Decimal>,
     pub bs: Option<u64>,
     pub ax: Option<String>,
-    pub ap: Option<f64>,
+    pub ap: Option<Decimal>,
     #[serde(rename = "as")]
     pub r#as: Option<u64>,
     pub c: Option<String>,
@@ -29,18 +29,18 @@ pub struct Quote {
 pub struct Trade {
     pub t: Option<Timestamp>,
     pub x: Option<String>,
-    pub p: Option<f64>,
+    pub p: Option<Decimal>,
     pub s: Option<u64>,
     pub c: Option<String>,
 }
 
 #[derive(Clone, Debug, Default, PartialEq, serde::Deserialize)]
 pub struct Greeks {
-    pub delta: Option<f64>,
-    pub gamma: Option<f64>,
-    pub rho: Option<f64>,
-    pub theta: Option<f64>,
-    pub vega: Option<f64>,
+    pub delta: Option<Decimal>,
+    pub gamma: Option<Decimal>,
+    pub rho: Option<Decimal>,
+    pub theta: Option<Decimal>,
+    pub vega: Option<Decimal>,
 }
 
 #[derive(Clone, Debug, Default, PartialEq, serde::Deserialize)]
@@ -52,5 +52,5 @@ pub struct Snapshot {
     pub dailyBar: Option<Bar>,
     pub prevDailyBar: Option<Bar>,
     pub greeks: Option<Greeks>,
-    pub impliedVolatility: Option<f64>,
+    pub impliedVolatility: Option<Decimal>,
 }
