+++ src/stocks/model.rs
@@ -1,10 +1,10 @@
-use crate::common::time::Timestamp;
+use crate::common::{decimal::Decimal, time::Timestamp};
 
 #[derive(Clone, Debug, Default, PartialEq, serde::Deserialize)]
 pub struct Auction {
     pub t: Option<Timestamp>,
     pub x: Option<String>,
-    pub p: Option<f64>,
+    pub p: Option<Decimal>,
     pub s: Option<u64>,
     pub c: Option<String>,
 }
@@ -19,23 +19,23 @@ pub struct DailyAuction {
 #[derive(Clone, Debug, Default, PartialEq, serde::Deserialize)]
 pub struct Bar {
     pub t: Option<Timestamp>,
-    pub o: Option<f64>,
-    pub h: Option<f64>,
-    pub l: Option<f64>,
-    pub c: Option<f64>,
+    pub o: Option<Decimal>,
+    pub h: Option<Decimal>,
+    pub l: Option<Decimal>,
+    pub c: Option<Decimal>,
     pub v: Option<u64>,
     pub n: Option<u64>,
-    pub vw: Option<f64>,
+    pub vw: Option<Decimal>,
 }
 
 #[derive(Clone, Debug, Default, PartialEq, serde::Deserialize)]
 pub struct Quote {
     pub t: Option<Timestamp>,
     pub bx: Option<String>,
-    pub bp: Option<f64>,
+    pub bp: Option<Decimal>,
     pub bs: Option<u64>,
     pub ax: Option<String>,
-    pub ap: Option<f64>,
+    pub ap: Option<Decimal>,
     #[serde(rename = "as")]
     pub r#as: Option<u64>,
     pub c: Option<Vec<String>>,
@@ -46,7 +46,7 @@ pub struct Quote {
 pub struct Trade {
     pub t: Option<Timestamp>,
     pub x: Option<String>,
-    pub p: Option<f64>,
+    pub p: Option<Decimal>,
     pub s: Option<u64>,
     pub i: Option<u64>,
     pub c: Option<Vec<String>>,
